Add files via upload
It is a dataset consisting of sounds collected from faucets and flushes used in WCs. Faucet data consists of 1, 2 and 3 faucet sounds and the number of these faucets and the recorded water sound according to 3 different flow levels. In addition, it is a dataset that includes sounds such as hand washing and flushing.